+++ src/io/mod.rs
@@ -197,6 +197,28 @@ pub trait Database: Sized {
                         .with(("price", DataType::float_interval(0., 50.))),
                 )
                 .build(),
+            TableBuilder::new()
+                .name("large_user_table")
+                .size(100000)
+                .schema(
+                    Schema::empty()
+                        .with(("id", DataType::integer_interval(0, 1000)))
+                        .with(("name", DataType::text()))
+                        .with((
+                            "age",
+                            DataType::optional(DataType::float_interval(0., 200.)),
+                        ))
+                        .with((
+                            "city",
+                            DataType::text_values([
+                                "Paris".into(),
+                                "New-York".into(),
+                                "Rome".into(),
+                            ]),
+                        ))
+                        .with(("income", DataType::float_interval(100.0, 200000.0))),
+                )
+                .build(),
         ]
     }